Force user name autocompletion for certain cmds
Certain commands in the Twitch chat require a user name to be entered
after the command. This commit enforces the auto completion to only
suggest user names and not to suggest any of the available emotes for
the user when providing the first argument.

This feature will work for the following commands at this moment of
writing.

/ban
/block
/mod
/purge
/timeout
/unban
/unblock
/unmod
/untimeout
/unvip
/user
/vip
/w