Update fetchRelationship to new Helix endpoint
/users/follows/ is no more, /channels/followed can be used to get the
relationship information. According to
https://dev.twitch.tv/docs/api/reference/#get-followed-channels the
response with both user_id and broadcaster_id is "...contains this
broadcaster if the user follows them. If not specified, the response
contains all broadcasters that the user follows" so the data array will
be either one entry or ALL entries. The total count will always be the
total number of followed channels.